+++ ChangeLog.md
@@ -5,9 +5,14 @@ Follow the principles in [keepachangelog.com](https://keepachangelog.com)!
 
 # v.Next (Current)
 
+### Fixed
+
+- Surround query filter's value with "/" when it is a number
+
+# [v0.2.0-beta.30](https://github.com/moxb/moxb/releases/tag/v0.2.0-beta.30) (2019-08-21)
+
 ### Added
-- [VS-139](https://goessntl.atlassian.net/browse/VS-139) Extend the NavControl interface with new APIs
-  (preparation for page enter and leave hooks)
+- Extend the NavControl interface with new APIs (preparation for page enter and leave hooks)
 
 # [v0.2.0-beta.25](https://github.com/moxb/moxb/releases/tag/v0.2.0-beta.25) (2019-06-27)